+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Version History
 v6.7.0
 ------
 
+* Fix telescope RA and Rotator display on DomeSummaryTable components. `<https://github.com/lsst-ts/LOVE-frontend/pull/685>`_
 * Add a button to log alarms from the AlarmsList and AlarmsTable components. `<https://github.com/lsst-ts/LOVE-frontend/pull/684>`_
 
 v6.6.0